+++ mcp_tools/cloud_container/cloud_and_container_security.py
@@ -5,101 +5,6 @@ from typing import Dict, Any
 def register_cloud_and_container_security_tools(mcp, hexstrike_client, logger):
 
     @mcp.tool()
-    def scout_suite_assessment(provider: str = "aws", profile: str = "default",
-                              report_dir: str = "/tmp/scout-suite", services: str = "",
-                              exceptions: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Scout Suite for multi-cloud security assessment.
-
-        Args:
-            provider: Cloud provider (aws, azure, gcp, aliyun, oci)
-            profile: AWS profile to use
-            report_dir: Directory to save reports
-            services: Specific services to assess
-            exceptions: Exceptions file path
-            additional_args: Additional Scout Suite arguments
-
-        Returns:
-            Multi-cloud security assessment results
-        """
-        data = {
-            "provider": provider,
-            "profile": profile,
-            "report_dir": report_dir,
-            "services": services,
-            "exceptions": exceptions,
-            "additional_args": additional_args
-        }
-        logger.info(f"☁️  Starting Scout Suite {provider} assessment")
-        result = hexstrike_client.safe_post("api/tools/scout-suite", data)
-        if result.get("success"):
-            logger.info(f"✅ Scout Suite assessment completed")
-        else:
-            logger.error(f"❌ Scout Suite assessment failed")
-        return result
-
-    @mcp.tool()
-    def cloudmapper_analysis(action: str = "collect", account: str = "",
-                            config: str = "config.json", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute CloudMapper for AWS network visualization and security analysis.
-
-        Args:
-            action: Action to perform (collect, prepare, webserver, find_admins, etc.)
-            account: AWS account to analyze
-            config: Configuration file path
-            additional_args: Additional CloudMapper arguments
-
-        Returns:
-            AWS network visualization and security analysis results
-        """
-        data = {
-            "action": action,
-            "account": account,
-            "config": config,
-            "additional_args": additional_args
-        }
-        logger.info(f"☁️  Starting CloudMapper {action}")
-        result = hexstrike_client.safe_post("api/tools/cloudmapper", data)
-        if result.get("success"):
-            logger.info(f"✅ CloudMapper {action} completed")
-        else:
-            logger.error(f"❌ CloudMapper {action} failed")
-        return result
-
-    @mcp.tool()
-    def pacu_exploitation(session_name: str = "hexstrike_session", modules: str = "",
-                         data_services: str = "", regions: str = "",
-                         additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Pacu for AWS exploitation framework.
-
-        Args:
-            session_name: Pacu session name
-            modules: Comma-separated list of modules to run
-            data_services: Data services to enumerate
-            regions: AWS regions to target
-            additional_args: Additional Pacu arguments
-
-        Returns:
-            AWS exploitation framework results
-        """
-        data = {
-            "session_name": session_name,
-            "modules": modules,
-            "data_services": data_services,
-            "regions": regions,
-            "additional_args": additional_args
-        }
-        logger.info(f"☁️  Starting Pacu AWS exploitation")
-        result = hexstrike_client.safe_post("api/tools/pacu", data)
-        if result.get("success"):
-            logger.info(f"✅ Pacu exploitation completed")
-        else:
-            logger.error(f"❌ Pacu exploitation failed")
-        return result
-
-    @mcp.tool()
     def kube_hunter_scan(target: str = "", remote: str = "", cidr: str = "",
                         interface: str = "", active: bool = False, report: str = "json",
                         additional_args: str = "") -> Dict[str, Any]:
