+++ mcp_tools/cloud_exploit/pacu.py
@@ -1,0 +1,37 @@
+# mcp_tools/cloud_exploit/pacu.py
+
+from typing import Dict, Any
+
+def register_pacu_tool(mcp, hexstrike_client, logger):
+    
+    @mcp.tool()
+    def pacu_exploitation(session_name: str = "hexstrike_session", modules: str = "",
+                         data_services: str = "", regions: str = "",
+                         additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Pacu for AWS exploitation framework.
+
+        Args:
+            session_name: Pacu session name
+            modules: Comma-separated list of modules to run
+            data_services: Data services to enumerate
+            regions: AWS regions to target
+            additional_args: Additional Pacu arguments
+
+        Returns:
+            AWS exploitation framework results
+        """
+        data = {
+            "session_name": session_name,
+            "modules": modules,
+            "data_services": data_services,
+            "regions": regions,
+            "additional_args": additional_args
+        }
+        logger.info(f"☁️  Starting Pacu AWS exploitation")
+        result = hexstrike_client.safe_post("api/tools/pacu", data)
+        if result.get("success"):
+            logger.info(f"✅ Pacu exploitation completed")
+        else:
+            logger.error(f"❌ Pacu exploitation failed")
+        return result
