+++ mcp_tools/cloud_visual/cloudmapper.py
@@ -1,0 +1,33 @@
+# mcp_tools/cloud_visual/cloudmapper.py
+
+from typing import Dict, Any
+
+def register_cloudmapper_tool(mcp, hexstrike_client, logger):
+    @mcp.tool()
+    def cloudmapper_analysis(action: str = "collect", account: str = "",
+                            config: str = "config.json", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute CloudMapper for AWS network visualization and security analysis.
+
+        Args:
+            action: Action to perform (collect, prepare, webserver, find_admins, etc.)
+            account: AWS account to analyze
+            config: Configuration file path
+            additional_args: Additional CloudMapper arguments
+
+        Returns:
+            AWS network visualization and security analysis results
+        """
+        data = {
+            "action": action,
+            "account": account,
+            "config": config,
+            "additional_args": additional_args
+        }
+        logger.info(f"☁️  Starting CloudMapper {action}")
+        result = hexstrike_client.safe_post("api/tools/cloudmapper", data)
+        if result.get("success"):
+            logger.info(f"✅ CloudMapper {action} completed")
+        else:
+            logger.error(f"❌ CloudMapper {action} failed")
+        return result
